Lab_4/Flask_Server/requirements.txt
requests~=2.32.3
pyjwt~=2.8.0
Flask~=3.0.3